app/requirements.txt
fastapi
uvicorn[standard]
pillow
opencv-python
numpy
face_recognition
